+++ lib/ui/components/drawer.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:printnotes/providers/navigation_provider.dart';
+import 'package:printnotes/providers/settings_provider.dart';
 
 import 'package:printnotes/ui/screens/archive/archive_screen.dart';
 import 'package:printnotes/ui/screens/trash/trash_screen.dart';
@@ -6,11 +8,12 @@ import 'package:printnotes/ui/screens/settings/settings_screen.dart';
 import 'package:printnotes/ui/screens/about/about_screen.dart';
 
 import 'package:printnotes/constants/constants.dart';
+import 'package:provider/provider.dart';
 
 class DrawerView extends StatelessWidget {
   const DrawerView({super.key});
 
-  void _navigateToScreen(BuildContext context, {Widget? screen}) {
+  void _navigateToScreen(BuildContext context, String path, {Widget? screen}) {
     final bool isDrawerPersistent = MediaQuery.sizeOf(context).width >= 800;
     // On desktop, drawer is a side menu, trying to pop it causes issues
     // this stops it.
@@ -19,9 +22,11 @@ class DrawerView extends StatelessWidget {
     }
 
     if (screen != null) {
+      context.read<NavigationProvider>().addToRouteHistory(path);
       Navigator.of(context).push(MaterialPageRoute(
         builder: (context) => screen,
       ));
+      context.read<NavigationProvider>().navigateBack();
     }
   }
 
@@ -72,37 +77,40 @@ class DrawerView extends StatelessWidget {
                 iconColor: Theme.of(context).colorScheme.secondary,
                 leading: const Icon(Icons.article_outlined),
                 title: const Text('All Notes'),
-                onTap: () => _navigateToScreen(context),
+                onTap: () => _navigateToScreen(
+                    context, context.read<SettingsProvider>().mainDir),
               ),
               const Opacity(opacity: 0.2, child: Divider()),
               ListTile(
                 iconColor: Theme.of(context).colorScheme.secondary,
                 leading: const Icon(Icons.archive_outlined),
                 title: const Text('Archive'),
-                onTap: () =>
-                    _navigateToScreen(context, screen: const ArchiveScreen()),
+                onTap: () => _navigateToScreen(
+                    context, context.read<SettingsProvider>().archivePath,
+                    screen: const ArchiveScreen()),
               ),
               ListTile(
                 iconColor: Theme.of(context).colorScheme.secondary,
                 leading: const Icon(Icons.delete_outlined),
                 title: const Text('Trash'),
-                onTap: () =>
-                    _navigateToScreen(context, screen: const DeletedScreen()),
+                onTap: () => _navigateToScreen(
+                    context, context.read<SettingsProvider>().trashPath,
+                    screen: const DeletedScreen()),
               ),
               const Opacity(opacity: 0.2, child: Divider()),
               ListTile(
                 iconColor: Theme.of(context).colorScheme.secondary,
                 leading: const Icon(Icons.settings_outlined),
                 title: const Text('Settings'),
-                onTap: () =>
-                    _navigateToScreen(context, screen: const SettingsScreen()),
+                onTap: () => _navigateToScreen(context, 'settingsScreen',
+                    screen: const SettingsScreen()),
               ),
               ListTile(
                 iconColor: Theme.of(context).colorScheme.secondary,
                 leading: const Icon(Icons.info_outlined),
                 title: const Text('About'),
-                onTap: () =>
-                    _navigateToScreen(context, screen: const AboutScreen()),
+                onTap: () => _navigateToScreen(context, 'aboutScreen',
+                    screen: const AboutScreen()),
               ),
             ],
           ),