+++ lib/ui/screens/trash/trash_screen.dart
@@ -5,7 +5,7 @@ import 'package:path/path.dart' as path;
 import 'package:provider/provider.dart';
 
 import 'package:printnotes/providers/settings_provider.dart';
-import 'package:printnotes/utils/handlers/item_navigation.dart';
+import 'package:printnotes/providers/navigation_provider.dart';
 import 'package:printnotes/utils/handlers/item_delete.dart';
 import 'package:printnotes/utils/storage_system.dart';
 
@@ -20,7 +20,6 @@ class _DeletedScreenState extends State<DeletedScreen> {
   List<FileSystemEntity> _deletedItems = [];
   String _currentPath = '';
   String _currentFolderName = 'Trash';
-  List<String> _folderHistory = [];
 
   @override
   void initState() {
@@ -38,7 +37,6 @@ class _DeletedScreenState extends State<DeletedScreen> {
       _currentPath = currentPath;
       _currentFolderName =
           _currentPath == deletePath ? 'Trash' : path.basename(currentPath);
-      _folderHistory = ItemNavHandler.folderHistory(deletePath);
     });
   }
 
@@ -90,11 +88,9 @@ class _DeletedScreenState extends State<DeletedScreen> {
       onTap: () {
         if (isDirectory) {
           _loadDeletedItems(item.path);
-          ItemNavHandler.addToFolderHistory(_currentPath,
-              dirHistory: _folderHistory);
+          context.read<NavigationProvider>().addToRouteHistory(_currentPath);
         } else {
-          ItemNavHandler.routeItemToPage(
-              context, item, () => _loadDeletedItems(_currentPath));
+          context.read<NavigationProvider>().routeItemToPage(context, item);
         }
       },
       onLongPress: () => _showBottomSheet(context, item),
@@ -124,14 +120,14 @@ class _DeletedScreenState extends State<DeletedScreen> {
         backgroundColor: Theme.of(context).colorScheme.primary,
         foregroundColor: Theme.of(context).colorScheme.onPrimary,
         title: Text(_currentFolderName),
-        leading: _folderHistory.length > 1
+        leading: context.read<NavigationProvider>().routeHistory.length > 2
             ? IconButton(
                 icon: const Icon(
                   Icons.arrow_back,
                 ),
                 onPressed: () {
                   setState(() => _loadDeletedItems(
-                      ItemNavHandler.navigateBack(dirHistory: _folderHistory)));
+                      context.read<NavigationProvider>().navigateBack()));
                 },
               )
             : null,