+++ lib/ui/widgets/speed_dial_fab.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_speed_dial/flutter_speed_dial.dart';
 
-import 'package:printnotes/providers/settings_provider.dart';
-import 'package:printnotes/utils/handlers/item_navigation.dart';
+import 'package:printnotes/providers/navigation_provider.dart';
 import 'package:printnotes/utils/handlers/item_create.dart';
 
-Widget speedDialFAB(BuildContext context, {required VoidCallback onLoadItems}) {
+Widget speedDialFAB(
+    BuildContext context, String currentFolder, Function loadItems) {
   return SpeedDial(
     icon: Icons.add,
     activeIcon: Icons.close,
@@ -17,25 +17,19 @@ Widget speedDialFAB(BuildContext context, {required VoidCallback onLoadItems}) {
         child: const Icon(Icons.create_new_folder_outlined),
         label: 'Create Folder',
         onTap: () => ItemCreationHandler.handleCreateNewFolder(
-          context,
-          context.read<SettingsProvider>().mainDir,
-          () => onLoadItems(),
-        ),
+            context, currentFolder, loadItems),
       ),
       SpeedDialChild(
         child: const Icon(Icons.note_add_outlined),
         label: 'Create Note',
         onTap: () => ItemCreationHandler.handleCreateNewNote(
-          context,
-          context.read<SettingsProvider>().mainDir,
-          () => onLoadItems(),
-        ),
+            context, currentFolder, loadItems),
       ),
       SpeedDialChild(
         child: const Icon(Icons.folder_copy),
         label: 'Open External File',
         onTap: () =>
-            ItemNavHandler.openExternalFile(context, () => onLoadItems()),
+            context.read<NavigationProvider>().openExternalFile(context),
       )
     ],
   );